Tokenizer.java finished Evaluator.java added ReversePolishNotation.java added Access modifier changed on every operation evaluate method Operand.java value data type changed from float to double Trigonometric classes' evaluate method body changed Syntax errors are handled Math errors (like Divide by Zero, Sqrt of negative number, tangent of 90 degree) are handled Model.java evaluate method renamed to solve, and finished

+++ app/src/main/java/com/example/calcumv20/Tokens/Operand/Operand.java
@@ -5,20 +5,20 @@ import androidx.annotation.NonNull;
 import com.example.calcumv20.Tokens.*;
 
 public class Operand extends Token {
-    private float value;
+    private double value;
 
-    public Operand(float value) {
+    public Operand(double value) {
         super(TokenType.OPERAND);
         this.value = value;
     }
 
-    public float getValue() {
+    public double getValue() {
         return value;
     }
 
     @NonNull
     @Override
     public String toString() {
-        return Float.toString(value);
+        return Double.toString(value);
     }
 }
